+++ CXGChat/ChatManager.hpp
@@ -32,9 +32,9 @@ public:
     
     virtual void OnMsg(PtlBase* ptl);
     
-    virtual void OnError(int errcode, char* msg);
+    virtual void OnError(long errcode, char* msg);
     
-    void setHost(const char *ip, int port);
+    void setHost(const char *ip, long port);
     void setUser(const char *uid,const char *rid,const char *token);
     void setController(ChatControl* listener);
